Struct alloy_signer_gcp::signer::GcpSigner

pub struct GcpSigner {
    client: GoogleApi<KeyManagementServiceClient<GoogleAuthMiddleware>>,
    key_name: String,
    chain_id: Option<ChainId>,
    pubkey: VerifyingKey,
    address: Address,
}

Google Cloud Platform Key Management Service (GCP KMS) Ethereum signer.

The GCP Signer passes signing requests to the cloud service. GCP KMS keys belong to a key ring, which is identified by a project ID, location, and key ring name. The key ring contains keys, which are identified by a key ID and version.

Because the public key is unknown, we retrieve it on instantiation of the signer. This means that the new function is async and must be called within some runtime.

Note that this wallet only supports asynchronous operations. Calling a non-asynchronous method will always return an error.

Examples

use alloy_signer::Signer;
use alloy_signer_gcp::{GcpKeyRingRef, GcpSigner, KeySpecifier};
use gcloud_sdk::{
    google::cloud::kms::v1::key_management_service_client::KeyManagementServiceClient,
    GoogleApi,
};

let project_id = std::env::var("GOOGLE_PROJECT_ID").expect("GOOGLE_PROJECT_ID");
let location = std::env::var("GOOGLE_LOCATION").expect("GOOGLE_LOCATION");
let keyring_name = std::env::var("GOOGLE_KEYRING").expect("GOOGLE_KEYRING");

let keyring = GcpKeyRingRef::new(&project_id, &location, &keyring_name);
let client = GoogleApi::from_function(
    KeyManagementServiceClient::new,
    "https://cloudkms.googleapis.com",
    None,
)
.await
.expect("Failed to create GCP KMS Client");

let key_name = "...";
let key_version = 1;
let key_specifier = KeySpecifier::new(keyring, key_name, key_version);
let chain_id = Some(1);
let signer = GcpSigner::new(client, key_specifier, chain_id).await.unwrap();

let message = vec![0, 1, 2, 3];

let sig = signer.sign_message(&message).await.unwrap();
assert_eq!(sig.recover_address_from_msg(message).unwrap(), signer.address());

Fields
client: GoogleApi<KeyManagementServiceClient<GoogleAuthMiddleware>>key_name: Stringchain_id: Option<ChainId>pubkey: VerifyingKeyaddress: Address
Implementations

impl GcpSigner

pub async fn new( client: GoogleApi<KeyManagementServiceClient<GoogleAuthMiddleware>>, key_specifier: KeySpecifier, chain_id: Option<ChainId>, ) -> Result<Self, GcpSignerError>

Instantiate a new signer from an existing Client, keyring reference, key ID, and version.

Retrieves the public key from GCP and calculates the Ethereum address.

pub async fn get_pubkey(&self) -> Result<VerifyingKey, GcpSignerError>

Fetch the pubkey associated with this signer’s key.

pub async fn sign_digest( &self, digest: &B256, ) -> Result<Signature, GcpSignerError>

Sign a digest with this signer’s key

async fn sign_digest_inner( &self, digest: &B256, ) -> Result<Signature, GcpSignerError>

Sign a digest with this signer’s key and add the eip155 v value corresponding to the input chain_id